requirements.txt
bcrypt
bs4
chromadb
docx2txt
google-generativeai
jinja2
langchain
langchain-community
langchain-google-genai
nest-asyncio
numpy
# openai
opencv-python
openpyxl
paddleocr
pandas
pymupdf
pypdf
pytesseract
python-docx
python-dotenv
pyyaml
sqlalchemy
streamlit
streamlit-authenticator==0.2.3
tesseract
tiktoken
unstructured[local-inference]
xlrd
